Remove old code and dependencies
