+++ src/lib.rs
@@ -1,301 +1,324 @@
 use egg::*;
-
+use indexmap::IndexMap;
+use rand::Rng;
+use rand_pcg::Pcg64;
 use std::{
     collections::{HashMap, HashSet},
-    fmt::{Debug, Display},
-    hash::Hash,
+    fmt::Display,
+    time::Duration,
 };
 
-pub trait Sample<T, C> {
-    fn get_random_vec(&mut self) -> Vec<C>;
+type Constant = i32;
+
+define_language! {
+    pub enum SimpleMath {
+        "+" = Add([Id; 2]),
+        "-" = Sub([Id; 2]),
+        "*" = Mul([Id; 2]),
+        "~" = Neg(Id),
+        // "/" = Div([Id; 2]),
+        Num(Constant),
+        Var(egg::Symbol),
+    }
 }
 
-pub trait Synthesizer<L, A>
-where
-    L: Language + 'static,
-    A: Analysis<L> + Clone,
-{
-    type CharacteristicVector: Hash + PartialEq + Eq + Clone + Debug + PartialOrd + Ord;
+#[derive(Default, Clone)]
+pub struct SynthAnalysis {
+    n_samples: usize,
+}
 
-    fn value_to_node(val: &Self::CharacteristicVector) -> Option<L>;
-    fn symbol_to_node(sym: Symbol) -> L;
-    fn node_to_symbol(node: &L) -> Option<Symbol>;
+impl Analysis<SimpleMath> for SynthAnalysis {
+    // doesnt need to be option
+    type Data = Vec<Constant>;
 
-    fn make_node(&mut self, egraph: &EGraph<L, A>) -> L;
-    fn initial_egraph(&mut self, params: &SynthesisParams<L, A>) -> EGraph<L, A>;
-    fn eval(
-        &mut self,
-        enode: &L,
-        egraph: &EGraph<L, A>,
-        values: &HashMap<Id, Self::CharacteristicVector>,
-    ) -> Self::CharacteristicVector;
-
-    fn instantiate(pattern: &Pattern<L>) -> RecExpr<L> {
-        let nodes: Vec<_> = pattern
-            .ast
-            .as_ref()
-            .iter()
-            .map(|n| match n {
-                ENodeOrVar::ENode(n) => n.clone(),
-                ENodeOrVar::Var(v) => {
-                    let s = v.to_string();
-                    assert!(s.starts_with('?'));
-                    Self::symbol_to_node(s[1..].into())
-                }
-            })
-            .collect();
-
-        RecExpr::from(nodes)
+    fn merge(&self, to: &mut Self::Data, from: Self::Data) -> bool {
+        // only do assertions on non-empty vecs
+        // there may be "bad" merges during minimization, that's ok
+        if !to.is_empty() && !from.is_empty() {
+            assert_eq!(to, &from);
+        }
+        false
     }
 
-    fn generalize(expr: &RecExpr<L>, map: &mut HashMap<Symbol, Var>) -> Pattern<L> {
-        let alpha = b"abcdefghijklmnopqrstuvwxyz";
-        let nodes: Vec<_> = expr
-            .as_ref()
-            .iter()
-            .map(|n| match Self::node_to_symbol(n) {
-                Some(sym) => {
-                    let var = if let Some(var) = map.get(&sym) {
-                        *var
-                    } else {
-                        let var = format!("?{}", alpha[map.len()] as char).parse().unwrap();
-                        map.insert(sym, var);
-                        var
-                    };
-                    ENodeOrVar::Var(var)
-                }
-                None => ENodeOrVar::ENode(n.clone()),
-            })
-            .collect();
-
-        Pattern::from(PatternAst::from(nodes))
+    fn make(egraph: &EGraph<SimpleMath, Self>, enode: &SimpleMath) -> Self::Data {
+        let x = |i: &Id| egraph[*i].data.iter().copied();
+        let params = &egraph.analysis;
+        match enode {
+            SimpleMath::Num(n) => (0..params.n_samples).map(|_| *n).collect(),
+            SimpleMath::Var(_) => vec![],
+            SimpleMath::Add([a, b]) => x(a).zip(x(b)).map(|(x, y)| x.wrapping_add(y)).collect(),
+            SimpleMath::Sub([a, b]) => x(a).zip(x(b)).map(|(x, y)| x.wrapping_sub(y)).collect(),
+            SimpleMath::Mul([a, b]) => x(a).zip(x(b)).map(|(x, y)| x.wrapping_mul(y)).collect(),
+            SimpleMath::Neg(a) => x(a).map(|x| -x).collect(),
+            // SimpleMath::Div([a, b]) => x(a).zip(x(b)).map(|(x, y)| x / y).collect(),
+        }
     }
 
-    fn minimize_equalities(
-        &mut self,
-        analysis: A,
-        equalities: &mut Vec<Equality<L, A>>,
-    ) -> Vec<Equality<L, A>> {
-        let mut removed = vec![];
-        'outer: while equalities.len() > 1 {
-            for (i, eq) in equalities.iter().enumerate() {
-                let other_eqs = equalities
-                    .iter()
-                    .enumerate()
-                    .filter(|(j, _)| i != *j)
-                    .flat_map(|(_, eq)| &eq.rewrites);
-                println!("minimizing");
-                let runner: Runner<L, A, ()> = Runner::new(analysis.clone())
-                    .with_expr(&Self::instantiate(&eq.lhs))
-                    .with_expr(&Self::instantiate(&eq.rhs))
-                    .with_node_limit(3000)
-                    .with_iter_limit(3)
-                    .run(other_eqs);
-                let are_same = |a, b| runner.egraph.find(a) == runner.egraph.find(b);
-                if are_same(runner.roots[0], runner.roots[1]) {
-                    println!("Removing {}", eq);
-                    removed.push(equalities.remove(i));
-                    continue 'outer;
+    fn modify(egraph: &mut EGraph<SimpleMath, Self>, id: Id) {
+        let cv = &egraph[id].data;
+        if cv.is_empty() {
+            return;
+        }
+        let first = cv[0];
+        if cv.iter().all(|x| *x == first) {
+            let added = egraph.add(SimpleMath::Num(first));
+            egraph.union(id, added);
+        }
+    }
+}
+
+fn generalize(expr: &RecExpr<SimpleMath>, map: &mut HashMap<Symbol, Var>) -> Pattern<SimpleMath> {
+    let alpha = b"abcdefghijklmnopqrstuvwxyz";
+    let nodes: Vec<_> = expr
+        .as_ref()
+        .iter()
+        .map(|n| match n {
+            SimpleMath::Var(sym) => {
+                let var = if let Some(var) = map.get(&sym) {
+                    *var
+                } else {
+                    let var = format!("?{}", alpha[map.len()] as char).parse().unwrap();
+                    map.insert(*sym, var);
+                    var
+                };
+                ENodeOrVar::Var(var)
+            }
+            n => ENodeOrVar::ENode(n.clone()),
+        })
+        .collect();
+
+    Pattern::from(PatternAst::from(nodes))
+}
+
+fn instantiate(pattern: &Pattern<SimpleMath>) -> RecExpr<SimpleMath> {
+    let nodes: Vec<_> = pattern
+        .ast
+        .as_ref()
+        .iter()
+        .map(|n| match n {
+            ENodeOrVar::ENode(n) => n.clone(),
+            ENodeOrVar::Var(v) => {
+                let s = v.to_string();
+                assert!(s.starts_with('?'));
+                SimpleMath::Var(s[1..].into())
+            }
+        })
+        .collect();
+
+    RecExpr::from(nodes)
+}
+
+fn minimize_equalities(
+    analysis: SynthAnalysis,
+    equalities: &mut Vec<Equality<SimpleMath, SynthAnalysis>>,
+) -> Vec<Equality<SimpleMath, SynthAnalysis>> {
+    let mut all_removed = vec![];
+
+    // dedup based on name
+    let mut set = HashSet::new();
+    equalities.retain(|eq| set.insert(eq.name.clone()));
+
+    // TODO we probably want some better heuristic on how general a rule is
+    // reversing the equalities puts the new ones first,
+    // since we want to remove them first
+    equalities.sort_by_key(|eq| {
+        let l = eq.lhs.ast.as_ref().len();
+        let r = eq.rhs.ast.as_ref().len();
+        (l.min(r), l.max(r))
+    });
+    equalities.reverse();
+
+    let mut granularity = equalities.len() / 2;
+    while granularity > 0 {
+        println!("Minimizing with granularity {}...", granularity);
+        let mut i = 0;
+        while i + granularity < equalities.len() {
+            let (before, tail) = equalities.split_at(i);
+            let (test, after) = tail.split_at(granularity);
+            i += granularity;
+
+            let mut runner: Runner<_, _, ()> = Runner::new(analysis.clone())
+                .with_node_limit(3000)
+                .with_iter_limit(5);
+
+            // Add the eqs to test in to the egraph
+            for eq in test {
+                runner = runner
+                    .with_expr(&instantiate(&eq.lhs))
+                    .with_expr(&instantiate(&eq.rhs));
+            }
+
+            let rewrites = before.iter().chain(after).flat_map(|eq| &eq.rewrites);
+            runner = runner.run(rewrites);
+
+            let mut to_remove = HashSet::new();
+            for (eq, roots) in test.iter().zip(runner.roots.chunks(2)) {
+                if runner.egraph.find(roots[0]) == runner.egraph.find(roots[1]) {
+                    to_remove.insert(eq.name.clone());
                 }
             }
 
-            break;
-        }
+            let (removed, kept) = equalities
+                .drain(..)
+                .partition(|eq| to_remove.contains(&eq.name));
+            *equalities = kept;
+            all_removed.extend(removed);
 
-        return removed;
-    }
-
-    fn run(&mut self, mut params: SynthesisParams<L, A>) -> Vec<Equality<L, A>> {
-        let mut egraph = self.initial_egraph(&params);
-        let mut values: HashMap<Id, Self::CharacteristicVector> = Default::default();
-
-        // initialize values map
-        for class in egraph.classes() {
-            if !values.contains_key(&class.id) {
-                // TODO passing the egraph here is a hack
-                let val = self.eval(&class.nodes[0], &egraph, &values);
-                values.insert(class.id, val);
+            if !to_remove.is_empty() {
+                println!("  Removed {} rules", to_remove.len());
+                for name in to_remove {
+                    println!("  removed {}", name);
+                }
             }
         }
 
-        for iter in 0..params.iterations {
-            println!("Iteration {}, known rules", iter);
-            for eq in &params.eqs {
+        granularity /= 2;
+    }
+
+    // reverse the list back
+    equalities.reverse();
+
+    return all_removed;
+}
+
+pub struct SynthParam {
+    pub rng: Pcg64,
+    pub n_iter: usize,
+    pub n_samples: usize,
+    pub variables: Vec<egg::Symbol>,
+    pub consts: Vec<Constant>,
+}
+
+impl SynthParam {
+    fn mk_egraph(&mut self) -> EGraph<SimpleMath, SynthAnalysis> {
+        let mut egraph = EGraph::new(SynthAnalysis {
+            n_samples: self.n_samples,
+        });
+        let rng = &mut self.rng;
+        for var in &self.variables {
+            let id = egraph.add(SimpleMath::Var(*var));
+            egraph[id].data = (0..self.n_samples).map(|_| rng.gen::<Constant>()).collect();
+        }
+        for c in &self.consts {
+            egraph.add(SimpleMath::Num(*c));
+        }
+        egraph
+    }
+
+    pub fn run(&mut self) -> Vec<Equality<SimpleMath, SynthAnalysis>> {
+        let mut equalities: Vec<Equality<SimpleMath, SynthAnalysis>> = vec![];
+        let mut eg = self.mk_egraph();
+        for iter in 0..self.n_iter {
+            println!(
+                "iter {} phase 1: adding ops over {} eclasses",
+                iter,
+                eg.number_of_classes()
+            );
+            let mut enodes_to_add = vec![];
+            for i in eg.classes() {
+                // enodes_to_add.push(SimpleMath::Neg(i.id));
+                for j in eg.classes() {
+                    enodes_to_add.push(SimpleMath::Add([i.id, j.id]));
+                    enodes_to_add.push(SimpleMath::Sub([i.id, j.id]));
+                    enodes_to_add.push(SimpleMath::Mul([i.id, j.id]));
+                }
+            }
+            for enode in enodes_to_add {
+                eg.add(enode);
+            }
+
+            println!(
+                "iter {} phase 2: running rules, n={}, e={}",
+                iter,
+                eg.total_size(),
+                eg.number_of_classes()
+            );
+            let rules = equalities.iter().flat_map(|eq| &eq.rewrites);
+            let runner: Runner<SimpleMath, SynthAnalysis, ()> =
+                Runner::new(eg.analysis.clone()).with_egraph(eg);
+            eg = runner
+                .with_time_limit(Duration::from_secs(20))
+                .with_node_limit(usize::MAX)
+                .with_iter_limit(3)
+                .run(rules)
+                .egraph;
+            println!(
+                "       phase 2: running rules, n={}, e={}",
+                eg.total_size(),
+                eg.number_of_classes()
+            );
+
+            println!("iter {} phase 3: discover rules", iter);
+            println!("       phase 3: grouping");
+            let mut by_cvec: IndexMap<&[Constant], Vec<Id>> = IndexMap::new();
+            for class in eg.classes() {
+                by_cvec.entry(&class.data).or_default().push(class.id);
+            }
+
+            println!("       phase 3: scanning {} groups", by_cvec.len());
+            let mut to_union = vec![];
+            let mut extract = Extractor::new(&eg, AstSize);
+
+            for ids in by_cvec.values() {
+                let cross = ids.iter().flat_map(|id1| {
+                    ids.iter()
+                        .filter_map(move |id2| if id1 > id2 { Some((id1, id2)) } else { None })
+                });
+
+                for (i, j) in cross {
+                    to_union.push((i.clone(), j.clone()));
+                    let (_cost1, expr1) = extract.find_best(i.clone());
+                    let (_cost2, expr2) = extract.find_best(j.clone());
+
+                    let names = &mut HashMap::default();
+                    let pat1 = generalize(&expr1, names);
+                    let pat2 = generalize(&expr2, names);
+
+                    if let Some(eq) = Equality::new(pat1, pat2) {
+                        equalities.push(eq);
+                    }
+                }
+            }
+
+            println!("       phase 3: performing {} unions", to_union.len());
+            for (i, j) in to_union {
+                eg.union(i, j);
+            }
+
+            let eq_len = equalities.len();
+            println!("iter {} phase 4: minimize {} rules", iter, eq_len);
+            minimize_equalities(eg.analysis.clone(), &mut equalities);
+            println!(
+                "iter {} phase 4: minimized {} to {} rules",
+                iter,
+                eq_len,
+                equalities.len()
+            );
+
+            println!("After iter {}, I know these equalities:", iter);
+            for eq in &equalities {
                 println!("  {}", eq);
             }
-
-            // update id_to_val based on the egraph modifications
-            values = values
-                .into_iter()
-                .map(|(id, val)| (egraph.find(id), val))
-                .collect();
-
-            for class in egraph.classes() {
-                if !values.contains_key(&class.id) {
-                    let node = class
-                        .iter()
-                        .find(|n| {
-                            n.children()
-                                .iter()
-                                .all(|id| values.contains_key(&egraph.find(*id)))
-                        })
-                        .unwrap_or_else(|| panic!("failed to find node for class {:?}", class));
-                    let val = self.eval(node, &egraph, &values);
-                    values.insert(class.id, val);
-                }
-            }
-
-            let mut ctr = 0;
-            // create the new nodes before adding to make sure they refer to existing nodes
-            let to_add: Vec<(L, Self::CharacteristicVector)> = (0..params.additions_per_iteration)
-                .map(|_| {
-                    let n = self.make_node(&egraph);
-                    println!("{}, evaling {:?}", ctr, n);
-                    let val = self.eval(&n, &egraph, &values);
-                    ctr = ctr + 1;
-                    (n, val)
-                })
-                .collect();
-
-            for (n, val) in to_add {
-                let n2 = Self::value_to_node(&val);
-                let id = egraph.add(n);
-                if let Some(n2) = n2 {
-                    let added = egraph.add(n2);
-                    egraph.union(id, added);
-                }
-                values.insert(id, val);
-            }
-            egraph.rebuild();
-
-            for class in egraph.classes() {
-                if !values.contains_key(&class.id) {
-                    let node = class
-                        .iter()
-                        .find(|n| {
-                            n.children()
-                                .iter()
-                                .all(|id| values.contains_key(&egraph.find(*id)))
-                        })
-                        .unwrap_or_else(|| panic!("failed to find node for class {:?}", class));
-                    let val = self.eval(node, &egraph, &values);
-                    values.insert(class.id, val);
-                }
-            }
-
-            // HACK with_egraph should create the runner
-            egraph = Runner::<L, A, ()>::new(egraph.analysis.clone())
-                .with_egraph(egraph)
-                .with_iter_limit(3)
-                .run(params.rewrites())
-                .egraph;
-            egraph.rebuild();
-
-            // update id_to_val based on the egraph modifications
-            values = values
-                .into_iter()
-                .map(|(id, sig)| (egraph.find(id), sig))
-                .collect();
-
-            // group things by value
-            let mut groups: HashMap<Self::CharacteristicVector, Vec<Id>> = Default::default();
-            for (id, val) in values.drain() {
-                assert_eq!(id, egraph.find(id));
-                groups.entry(val).or_default().push(egraph.find(id));
-            }
-            assert!(groups.len() <= egraph.number_of_classes());
-
-            let mut ext = Extractor::new(&egraph, AstSize);
-            for ids in groups.values_mut() {
-                ids.sort();
-                ids.dedup();
-                let len = ids.len();
-
-                for i in 0..len - 1 {
-                    let j = i + 1;
-                    // for j in i + 1..len {
-                    let var_map = &mut HashMap::default();
-                    let pat1 = Self::generalize(&ext.find_best(ids[i]).1, var_map);
-                    let pat2 = Self::generalize(&ext.find_best(ids[j]).1, var_map);
-                    if let Some(eq) = Equality::new(pat1, pat2) {
-                        println!("Learning {}", eq);
-                        params.eqs.push(eq);
-                    }
-                    // }
-                }
-            }
-
-            for (val, ids) in groups {
-                for &id in &ids {
-                    assert_eq!(id, egraph.find(id));
-                }
-                if ids.len() > 1 {
-                    println!("Unioning {:?}", ids)
-                }
-                let id = fold1(ids, |x, y| egraph.union(x, y).0);
-                let old = values.insert(egraph.find(id), val);
-                assert!(old.is_none());
-            }
-            egraph.rebuild();
-            // assert_eq!(values.len(), egraph.number_of_classes(), "here2");
-
-            println!("Minimizing {} equalities...", params.eqs.len());
-            self.minimize_equalities(egraph.analysis.clone(), &mut params.eqs);
-            println!("Minimized to {} equalities", params.eqs.len());
         }
-
-        println!("Final equalities");
-        for eq in &params.eqs {
-            println!("  {}", eq);
-        }
-        params.eqs
+        equalities
     }
-}
-
-fn fold1<T, F>(iter: impl IntoIterator<Item = T>, f: F) -> T
-where
-    F: FnMut(T, T) -> T,
-{
-    let mut iter = iter.into_iter();
-    let first = iter.next().expect("can't be empty");
-    iter.fold(first, f)
-}
-
-pub struct SynthesisParams<L, A> {
-    pub iterations: usize,
-    pub additions_per_iteration: usize,
-    pub eqs: Vec<Equality<L, A>>,
-}
-
-impl<L, A> SynthesisParams<L, A> {
-    fn rewrites(&self) -> impl Iterator<Item = &Rewrite<L, A>> {
-        self.eqs.iter().flat_map(|eq| &eq.rewrites)
-    }
-}
-
-#[derive(Debug, PartialEq, Eq, Hash)]
-struct Signature<L> {
-    samples: Vec<L>,
 }
 
 pub struct Equality<L, A> {
     pub lhs: Pattern<L>,
     pub rhs: Pattern<L>,
-    rewrites: Vec<egg::Rewrite<L, A>>,
+    pub name: String,
+    pub rewrites: Vec<egg::Rewrite<L, A>>,
 }
 
 impl<L: Language, A> Display for Equality<L, A> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self.rewrites.len() {
-            1 => write!(f, "{}", self.rewrites[0].long_name()),
-            2 => write!(f, "{} <=> {}", self.lhs, self.rhs),
-            n => panic!("unexpected len {}", n),
-        }
+        write!(f, "{}", self.name)
     }
 }
 
 impl<L: Language + 'static, A: Analysis<L>> Equality<L, A> {
-    pub fn new(lhs: Pattern<L>, rhs: Pattern<L>) -> Option<Self> {
+    fn new(lhs: Pattern<L>, rhs: Pattern<L>) -> Option<Self> {
         let mut rewrites = vec![];
 
         let name = format!("{} => {}", lhs, rhs);
@@ -311,7 +334,26 @@ impl<L: Language + 'static, A: Analysis<L>> Equality<L, A> {
         if rewrites.is_empty() {
             None
         } else {
-            Some(Self { rewrites, lhs, rhs })
+            let name = match rewrites.len() {
+                1 => format!("{}", rewrites[0].long_name()),
+                2 => {
+                    // canonicalize the name, as we use it for dedup
+                    let l_str = format!("{}", lhs);
+                    let r_str = format!("{}", rhs);
+                    if l_str < r_str {
+                        format!("{} <=> {}", l_str, r_str)
+                    } else {
+                        format!("{} <=> {}", r_str, l_str)
+                    }
+                }
+                n => panic!("unexpected len {}", n),
+            };
+            Some(Self {
+                rewrites,
+                lhs,
+                rhs,
+                name,
+            })
         }
     }
 }
